pyproject.toml
[tool.poetry]
name = "odd-collector-gcp"
version = "0.1.0"
description = "ODD Collector for GCP"
license = "Apache-2.0"
authors = ["Open Data Discovery <pypi@opendatadiscovery.org>"]
homepage = "https://github.com/opendatadiscovery/odd-collector-gcp"
repository = "https://github.com/opendatadiscovery/odd-collector-gcp"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
oddrn-generator = "^0.1.34"
odd-collector-sdk = "^0.2.1"
google-cloud-bigquery = "^3.1.0"
black = "^22.3.0"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
pytest = "^6.2.5"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
